Use BaseEntityData in PlayerData; Move read_position and read_velocity tests
